+++ internal/conf/write.go
@@ -25,7 +25,6 @@ func Write(config models.Conf) {
 	viper.Set("IFACES", config.Ifaces)
 	viper.Set("TIMEOUT", config.Timeout)
 	viper.Set("TRIM_HIST", config.TrimHist)
-	viper.Set("HIST_IN_DB", config.HistInDB)
 	viper.Set("SHOUTRRR_URL", config.ShoutURL)
 
 	viper.Set("USE_DB", config.UseDB)